moved files from main to other files
**lreywen**
- added sendas
**main**
- removed br_main & sendas
**revx**
- added br_main
