+++ src/main.rs
@@ -7,7 +7,7 @@ mod lib {
 }
 
 use crate::lib::{
-    message::{RMessage, RMessagePayload, Masquerade},
+    message::RMessage,
     br::BrConf,
     auth::Auth,
 };
@@ -81,7 +81,6 @@ async fn main()  {
 }
 
 // establishes websocket connection
-
 pub async fn websocket(url: String, authen: Auth, br: BrConf) {
 
      let (ws_stream, _response) = connect_async(url).await.expect("Failed to connect");
@@ -103,7 +102,6 @@ pub async fn websocket(url: String, authen: Auth, br: BrConf) {
 }
 
 
-
 // websocket main
 // imports messages, cleans them and sends to 
 // bridge and message processing
@@ -117,7 +115,6 @@ pub async fn newmain(authen: Auth, out: String, br: BrConf) {
         Ok(_) => print!("")
 
     };
-
     
     let message = inval_message;
     
@@ -127,77 +124,6 @@ pub async fn newmain(authen: Auth, out: String, br: BrConf) {
     if br.enabled == true {
         br_main(authen.clone(), inval2.unwrap(), br).await;
     };
-
-}
-// main for bridge, WIP
-pub async fn br_main(auth: Auth, input_message: RMessage, br: BrConf) {
-
-
-    let (chan1, chan2) = (br.channel_1, br.channel_2);
-    // removing feedback loop
-    if input_message.author == auth.bot_id && input_message.masquerade != None {
-        return
-    };
-
-
-    // channel switch
-    let mut chan_rec = String::new();
-    if input_message.channel == chan1 {
-       chan_rec = chan2;
-    }else if input_message.channel == chan2 {
-       chan_rec = chan1;
-    };
-
-
-    
-    let mut message = input_message.clone();
-    
-    message.channel = chan_rec;
-
-    let mut br_masq = Masquerade {
-        name: None,
-        avatar: None,
-        colour: None
-    };
-   
-    // masq switch - if user has no masquerade: pull from user info API
-    // else - port over masquerade details 
-    if input_message.masquerade == None {
-
-        // API get masq
-        
-        let user = rev_user(auth.clone(), input_message.author.clone()).await.expect("failed to GET user details");
-
-        let pfplink = user.avatar.unwrap().id;
-
-        let pfp = format!("https://autumn.revolt.chat/avatars/{pfplink}");
-
-        br_masq = Masquerade {
-            name: Some(user.username),
-            avatar: Some(pfp),
-            colour: None
-        };
-        
-    }else {
-        
-        // translate masq
-        br_masq = Masquerade {
-            name: message.masquerade.as_ref().unwrap().name.clone(),
-            avatar: message.masquerade.as_ref().unwrap().avatar.clone(),
-            colour: None
-        };  
-
-    };
-
-    // message for rev_send
-    let payload = RMessagePayload {
-        content: message.content.clone(),
-        attachments: None,
-        replies: rev_convert_reply(input_message.replies).await,
-        masquerade: Some(br_masq),
-    };
-
-    rev_send(auth, message, payload).await;
 
 }
 
@@ -232,48 +158,4 @@ pub async fn message_process(data: Auth, message_in: RMessage) {
         _ => return
     };
 
-}
-
-// masq wrapper for rev_send
-
-pub async fn sendas(auth: Auth, message: RMessage, content_vec: Vec<&str>) {
-
-    if content_vec.len() < 3 {
-        send(auth, message, "invalid use of sendas".to_string()).await;
-        return
-    };
-    //let from = message._id.clone();
-    let masq = content_vec[1];
-    let mut content = String::new();
-    //content = "placeholder".to_string();
-
-    let link = match masq {
-        "bingus"    | "cheese"  | "dad" | 
-        "deeznuts"  |  "insert" | "joe_biden" | 
-        "valence"   | "walter"  | "woof" => format!("https://toastxc.xyz/TXCS/{masq}.jpg"),
-        _ => format!("https://toastxc.xyz/TXCS/default.png")
-    };
-
-    for x in 0..content_vec.len() -2 {
-        content += &format!(" {}", content_vec[x + 2]);
-    };
-
-    let masq_s = Masquerade {
-        name: Some(masq.to_string()),
-        avatar: Some(link),
-        colour: None,
-    };
-
- 
-    let replier = rev_convert_reply(message.replies.clone()).await;
-
-    let returner = RMessagePayload {
-          content: Some(content),
-          replies: replier,
-          attachments: None,
-          masquerade: Some(masq_s)
-    };
-
-    rev_send(auth.clone(), message.clone(), returner).await;
-    rev_del(auth.clone(), message.clone()).await;
 }
