+++ src/rev_x.rs
@@ -1,10 +1,14 @@
 // structs
 
 use crate::lib::{
-    message::{RMessage, RMessagePayload, RReplies},
+    message::{RMessage, RMessagePayload, RReplies, Masquerade},
     auth::Auth,
-    user::RUserFetch
+    user::RUserFetch,
+    br::BrConf,
+
 };
+
+
 
 // dep
 use rand::Rng;
@@ -152,4 +156,74 @@ pub async fn rev_convert_reply(input: Option<Vec<String>>) -> Option<Vec<RReplie
 
 }
 
+pub async fn br_main(auth: Auth, input_message: RMessage, br: BrConf) {
+
+
+    let (chan1, chan2) = (br.channel_1, br.channel_2);
+    // removing feedback loop
+    if input_message.author == auth.bot_id && input_message.masquerade != None {
+        return
+    };
+
+
+    // channel switch
+    let mut chan_rec = String::new();
+    if input_message.channel == chan1 {
+       chan_rec = chan2;
+    }else if input_message.channel == chan2 {
+       chan_rec = chan1;
+    };
+
+
+    
+    let mut message = input_message.clone();
+    
+    message.channel = chan_rec;
+
+    let mut br_masq = Masquerade {
+        name: None,
+        avatar: None,
+        colour: None
+    };
+   
+    // masq switch - if user has no masquerade: pull from user info API
+    // else - port over masquerade details 
+    if input_message.masquerade == None {
+
+        // API get masq
+        
+        let user = rev_user(auth.clone(), input_message.author.clone()).await.expect("failed to GET user details");
+
+        let pfplink = user.avatar.unwrap().id;
+
+        let pfp = format!("https://autumn.revolt.chat/avatars/{pfplink}");
+
+        br_masq = Masquerade {
+            name: Some(user.username),
+            avatar: Some(pfp),
+            colour: None
+        };
+        
+    }else {
+        
+        // translate masq
+        br_masq = Masquerade {
+            name: message.masquerade.as_ref().unwrap().name.clone(),
+            avatar: message.masquerade.as_ref().unwrap().avatar.clone(),
+            colour: None
+        };  
+
+    };
+
+    // message for rev_send
+    let payload = RMessagePayload {
+        content: message.content.clone(),
+        attachments: None,
+        replies: rev_convert_reply(input_message.replies).await,
+        masquerade: Some(br_masq),
+    };
+
+    rev_send(auth, message, payload).await;
+
+}
